bug fix : dans m_envoi, correction de la maj de la liste chainee des cases libres et de l'ecriture de length dans les messages envoyes

+++ m_file.h
@@ -30,7 +30,7 @@ typedef struct mon_message{
 } mon_message;
 
 typedef struct header{
-	int max_length_message;
+	size_t max_length_message;
 	int pipe_capacity;
 	int first_occupied;
 	int last_occupied;
